Remove all uses of Gc::new_from_layout.
Mostly these are very simple changes. The only slightly -- but, fortunately, not
very -- tricky case is SOMStack. To minimise changes, I have chosen to use a
fixed size array *within* `SOMStack`. Because upvars use interior pointers to
the SOM stack, this design is fine, but it does impose a requirement (which we
happened to already meet) that `SOMStack` is only ever used as `GC<SOMStack>`.